requirements.txt
azure-storage-blob>=12.0.0
openai>=1.0.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
llama-index-core>=0.10.0
llama-index>=0.10.0
llama-index-vector-stores-postgres>=0.1.1
psycopg2-binary>=2.9.9
python-multipart>=0.0.5
typing-extensions>=4.0.0
pypdf>=3.0.0
docx2txt>=0.8

# Testing dependencies
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.20.0,<0.22.0
httpx>=0.24.0
coverage>=6.5.0,<7.0.0
pytest-cov>=4.0.0,<5.0.0 